# contracts/p5-reporting-decommission/openapi.yaml
openapi: 3.1.0
info:
  title: HRMS Phase 5 – Reporting / Integration, reference-data ownership, Forms + Oracle decommission
  version: 5.0.0-contract
  description: |
    Frozen API surface for Phase 5 (CUTOVER_PLAN.md §9, COMPONENT_MAPPING.md §9 rows
    `PKG_REPORTING` / `PKG_INTEGRATION`, MODERNIZATION_BLUEPRINT.md §8 Reporting row,
    TEST_STRATEGY.md §5 row 5). Written BEFORE any implementation exists; the backend
    node implements exactly this file, the frontend node codes against it.

    Target architecture (non-negotiable, MODERNIZATION_BLUEPRINT.md §9 decisions A/B):
    * Spring Boot (Java 21) on PostgreSQL + React. **No PL/SQL, PL/pgSQL, triggers or
      .pll in the target.** `PKG_REPORTING`, `PKG_INTEGRATION`, the six `VW_*` views and
      `HRMS_ADMIN.fmb` / `HRMS_REPORTS.fmb` are read-only characterization references
      (GOLDEN-ORACLE MODE = OFF: no Oracle runtime exists; expectations are recorded
      fixtures, `legacy_source=recorded`).
    * Three Java modules are introduced (ARCH-01/02, one owner per table):
      - `reporting-module` – **owns no table**. Every report is a read-only Spring Data
        JDBC projection (plain SQL, `WITH RECURSIVE` for the hierarchy) over tables owned
        by the P0–P4 modules, executed on the PostgreSQL read replica when one exists.
        The six `VW_*` views are **not** recreated on PostgreSQL; the report SQL is the
        `tests/reconciliation/pg/*.sql` transcription (Level 3 pins it).
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin-roles]
      operationId: adminSearchUserAccounts
      summary: User accounts with their roles (paged). Sorted by `username`.
      x-preauthorize: hasAuthority('ADMIN:VIEW')
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 100 }
          description: "Case-insensitive contains on `username`, employee number, first or last name."
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/AccountStatus' }
        - name: roleId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: locked
          in: query
          required: false
          schema: { type: boolean }
          description: "`true` → `locked_until > now()` only."
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Page of accounts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAccountPage' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
    get:
      tags: [admin-roles]
      operationId: adminGetUserAccount
      summary: One account with roles and effective authorities
      x-preauthorize: hasAuthority('ADMIN:VIEW')
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAccount' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
  /api/admin/users/{userId}/roles:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
    put:
      tags: [admin-roles]
      operationId: adminReplaceUserRoles
      summary: Replace the account's role set
      description: |
        Idempotent set replacement on `user_roles` (`granted_by = jwt.sub` on inserted rows).
        Rules: `roleIds` non-empty, distinct, every id exists (`-20807`
        `field=roleIds`); the caller may not change **their own** account
        (`jwt.sub == userId`, `-20804`); **least privilege** – every authority granted by the
        new set that the target does not already hold must be held by the caller (`-20806`);
        **last admin** – the result must leave at least one `ACTIVE` account with
        `ADMIN:EDIT` (`-20805`). Because the invariant spans *all* accounts, every mutation
        that can lower the `ADMIN:EDIT` holder count (`PUT …/roles`, `PUT …/status`,
        `PUT /api/admin/roles/{roleId}`) first takes the same transaction-scoped
        `pg_advisory_xact_lock(hashtext('hrms.admin_edit_guard'))`, then applies the write,
        then re-counts `ACTIVE` accounts holding `ADMIN:EDIT` and rolls back with `-20805`
        if zero; row locks alone are insufficient. Then all sessions of the target are revoked
        (`SessionRevoker`): the target's next `/api/auth/refresh` returns `401 TOKEN_INVALID`
        and they must log in again; `UserAccount.authorities` shows the new effective set.
      x-preauthorize: hasAuthority('ADMIN:EDIT')
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRolesRequest' }
      responses:
        '200':
          description: Replaced
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAccountWriteResult' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '422':
          $ref: '#/components/responses/RoleRuleViolation'
  /api/admin/users/{userId}/status:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
    put:
      tags: [admin-roles]
      operationId: adminSetUserAccountStatus
      summary: Enable / disable the account (`user_accounts.status`)
      description: |
        `DISABLED` revokes all sessions and makes `/api/auth/login` answer `401 TOKEN_INVALID`
        family (P0 `ACCOUNT_DISABLED`); `ACTIVE` re-enables and clears `locked_until` /
        `failed_attempts` (admin unlock). Self (`-20804`) and last-admin (`-20805`) rules
        apply. `modified_by = jwt.sub`.
      x-preauthorize: hasAuthority('ADMIN:EDIT')
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserStatusRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAccountWriteResult' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '422':
          $ref: '#/components/responses/RoleRuleViolation'

  # ---------------------------------------------------------------------------
  # admin – audit log search
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: P0 access token; claims `sub` (user-account id, string), `empId`, `authorities[]`.
  parameters:
...
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      in: query
      name: size
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    UserIdPath:
      in: path
      name: userId
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
  responses:
...
    ValidationFailed:
      description: Bean Validation / unsupported query value (`validation-schema.json` module `p5-reporting-decommission`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: VALIDATION_FAILED, message: asOf must be a date (YYYY-MM-DD), field: asOf, traceId: 4f3c… }
    Unauthorized:
      description: Missing / invalid / revoked token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: TOKEN_INVALID, message: Access token is missing or invalid, traceId: 4f3c… }
    Forbidden:
      description: "`@PreAuthorize` denied (`REPORTS:VIEW`, `PAYROLL:VIEW`, `ADMIN:VIEW`, `ADMIN:EDIT`, `LEAVE:ADMIN`, `PAYROLL:APPROVE`, `PAYROLL:EDIT`, `EMPLOYEE:VIEW`)"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: FORBIDDEN, message: Access denied, traceId: 4f3c… }
...
    UserNotFound:
      description: "`USER_NOT_FOUND`"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: USER_NOT_FOUND, message: 'User account not found: 42', traceId: 4f3c… }
...
    RoleRuleViolation:
      description: "`-20802` seeded role read-only, `-20804` self-modification, `-20805` last `ADMIN:EDIT` account, `-20806` privilege exceeds caller, `-20807` unknown role id"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: '-20805', message: 'Cannot remove the last active account holding ADMIN:EDIT', field: roleIds, traceId: 4f3c… }
...
  schemas:
    ApiError:
      type: object
      required: [code, message, traceId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: Legacy `-20xxx` as a string, or a framework code (`error-codes.md`).
        message: { type: string }
        field:
          type: [string, 'null']
          description: Offending request field (body property, query parameter, multipart part or `line N`) when attributable.
        traceId:
          type: string
          description: Correlates with `error_log.trace_id` / `X-Trace-Id`.
        details:
          type: array
          description: All violations for `VALIDATION_FAILED` / `-20704`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    Authority:
      type: string
      pattern: '^(PAYROLL|EMPLOYEE|LEAVE|ADMIN|REPORTS|PERFORMANCE):(VIEW|EDIT|APPROVE|CREATE|ADMIN|VIEW_ALL)$'
      description: |
        `role_permissions.authority` (`chk_rp_authority` as widened by V4). The assignable
        vocabulary served by `GET /api/admin/authorities` is exactly the set the backend
        references in `@PreAuthorize`: the P0 table `PAYROLL|EMPLOYEE|LEAVE|ADMIN|REPORTS`
        × `VIEW|EDIT|APPROVE|CREATE` plus `LEAVE:ADMIN`, `LEAVE:VIEW_ALL` and the
        `PERFORMANCE:*` actions. Values matching the pattern but absent from that set are
        `VALIDATION_FAILED` (`field=permissions`).
    AccountStatus:
      type: string
      enum: [ACTIVE, DISABLED]
      description: "`user_accounts.status` (`chk_ua_status`)."
...
    UserRoleGrant:
      type: object
      required: [roleId, roleCode, roleName, grantedBy, grantedDate]
      properties:
        roleId: { type: integer }
        roleCode: { type: string }
        roleName: { type: string }
        grantedBy: { type: string, description: "`user_roles.granted_by` – `jwt.sub` of the grantor or `SYSTEM`." }
        grantedDate: { type: string, format: date-time }
    UserAccount:
      type: object
      required: [userId, empId, empNumber, fullName, username, status, locked, mustChangePassword, roles, authorities, createdBy, createdDate]
      properties:
        userId: { type: integer, format: int64, description: "= `jwt.sub` of that user." }
        empId: { type: integer, format: int64 }
        empNumber: { type: string }
        fullName: { type: string }
        username: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        locked: { type: boolean, description: "`locked_until > now()`." }
        lockedUntil: { type: [string, 'null'], format: date-time }
        failedAttempts: { type: integer, minimum: 0 }
        mustChangePassword: { type: boolean }
        passwordChangedAt: { type: [string, 'null'], format: date-time }
        roles:
          type: array
          items: { $ref: '#/components/schemas/UserRoleGrant' }
        authorities:
          type: array
          description: "Effective set = `UserAccountRepository.authorities(userId)` (distinct union over roles), sorted."
          items: { $ref: '#/components/schemas/Authority' }
        createdBy: { type: string }
        createdDate: { type: string, format: date-time }
        modifiedBy: { type: [string, 'null'] }
        modifiedDate: { type: [string, 'null'], format: date-time }
    UserAccountPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/UserAccount' }
        page: { $ref: '#/components/schemas/PageMeta' }
    UserRolesRequest:
      type: object
      required: [roleIds]
      additionalProperties: false
      properties:
        roleIds:
          type: array
          minItems: 1
          uniqueItems: true
          description: "Bounded only by the set of existing roles and distinctness; no arbitrary cap."
          items: { type: integer, minimum: 1 }
    UserStatusRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { $ref: '#/components/schemas/AccountStatus' }
        reason: { type: [string, 'null'], maxLength: 200, description: "Copied into the `AUDIT_LOG` row's `new_values.reason`; never stored on `user_accounts`." }
    UserAccountWriteResult:
      allOf:
        - $ref: '#/components/schemas/UserAccount'
        - type: object
          required: [sessionsRevoked]
          properties:
            sessionsRevoked: { type: integer, minimum: 0 }

    # --- reports ---------------------------------------------------------------
...